Before you start navigation
